usecase/agi-agent-application/Miney/requirements.txt
python-dotenv
streamlit
streamlit-lottie
PyPDF2
streamlit_lottie
beautifulsoup4==4.12.3
bs4==0.0.2
dataclasses-json==0.6.7
datasets==3.4.1
faiss-cpu==1.10.0
kobert-tokenizer @ git+https://github.com/SKTBrain/KoBERT.git@5c46b1c68e4755b54879431bd302db621f4d2f47#subdirectory=kobert_hf
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
numpy==1.26.4
openai==1.66.3
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pydantic==2.8.2
pydantic-settings==2.8.1
pydantic_core==2.20.1
requests==2.32.3
requests-futures==1.0.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.13.0
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==2.19.2
setuptools==75.6.0
sympy==1.13.1
tokenizers==0.21.1
torch==2.5.1
torchmetrics==1.7.0
transformers==4.50.3
typing==3.7.4.3

